requirements.txt
stripe
selenium
Flask-API
mercadopago
flask-marshmallow
Flask-Mail
Flask-Session
google-api-python-client
firebase-admin
google-cloud-firestore
bottlenose
lxml
python-dateutil
python-amazon-simple-product-api
requests
flask-cors
Flask-SSLify
numpy
psycopg2-binary
SQLAlchemy
boto3
apscheduler
flask-cors
flask-restful
google-cloud-translate
googletrans
pymongo
dnspython
csvalidate
Flask-CSV
beautifulsoup4
zipfile36